HEP 10: Disable parallel tool calling.
It isn't correctly handled. This change does tweak it so that the openai
handler supports it (prints each tool call separately). However, the
caller logic that contructs the message history does not assemble the
messages correctly so it's disabled for now.

It's also not entirely clear yet how useful it is for CLI use cases.