HEP 18: Add `/asset-listen` to wait for changes to an asset.
`/asset-listen <name> [<cursor>]` blocks the repl until the asset has
been modified. On a change, new data info including data-url, data-size,
operation, metadata-url (if available), and new cursor are added to the
conversation for the AI.

The cursor is important to ensure that follow up calls to `/asset-listen`
do not miss a change event.

While a bit complex, the AI can be instructed to follow and print all
changes to an asset. For example, using gpt-4.1, this works:

> !hai listen to changes to asset `test`. When a change is detected,
> from the asset-change info in the convo, invoke python to download the
> data-url and print it. lastly, re-invoke !hai to repeat this process
> (/asset-listen called again with the new cursor, invoke python, ....).
> in other words, you should return 3 commands: /asset-listen ...,
> !py ..., !hai ...