HEP 10: Set file ext for temp files based on content_type.
Some editors (e.g. vim) use the extension to decide how to syntax
highlight a file. Previously, if an asset_name didn't include an
extension, the temp file for local editing wouldn't have an extension.
This change will add an extension if the temp file has a content_type
that is associated with an extension (text/markdown -> md) or defaults
to md if content_type isn't set and it begins with "# ".