HEP 10: /task: Set Konsole tab title to task name.
Unfortunately, Konsole uses a special escape code for setting tab titles
rather than applying a window title to a specific tab as other terminals
do. This change add supports for tab titles. Unfortunately, it's not
possible to reset the tab title to the previous value on /task-end, so
instead the default Konsole tab title is used.