HEP 10: Add experimental !pyuv tool.
A tool to run a python script using `uv run`. The advantage is that
script dependencies are automatically installed. These are deps defined
at the top of the file as follows:

    # /// script
    # dependencies = [
    # ...
    # ]
    # ///

This is an improvement over !'uv run --with dep' since the notation is
shorter and it doesn't require specifying the dependency upfront.