HEP 10: !fn-py: Do not print return value.
The return value hasn't been of use. So far, the useful values from the
AI-defined functions have been printed, which naturally adds them to the
chat history.

Also, removes `JsonCompatible` from the !fn-py generated-fn signature.
While it was an earlier consideration, it was phased out in favor of
using a native Python expression. Likewise, !fn-sh receives a native
shell value or expression.