HEP 30: Add `/std new-day-alert` stdlib fn.
Useful for tasks like ken/calendar where a user may return to the chat a
day later and it's important the AI be aware that the day has changed.

The check is triggered whenever a user inputs a new command. If it's a
new day, a user message indicating as such is added before their new
input.